requirements.txt
torch
torchvision
streamlit
pillow
requests